+++ DTTableViewManager/TableViewFactory.swift
@@ -41,30 +41,9 @@ class TableViewFactory
         self.tableView = tableView
     }
     
-    private func mappingForViewType(type: ViewType,modelTypeMirror: _MirrorType) -> ViewModelMapping?
-    {
-        let adjustedModelTypeMirror = RuntimeHelper.classClusterReflectionFromMirrorType(modelTypeMirror)
-        return self.mappings.filter({ (mapping) -> Bool in
-            return mapping.viewType == type && mapping.modelTypeMirror.summary == adjustedModelTypeMirror.summary
-        }).first
-    }
-    
-    private func addMappingForViewType<T:ModelTransfer>(type: ViewType, viewClass : T.Type)
-    {
-        if self.mappingForViewType(type, modelTypeMirror: _reflect(T.ModelType.self)) == nil
-        {
-            self.mappings.append(ViewModelMapping(viewType : type,
-                viewTypeMirror : _reflect(T),
-                modelTypeMirror: _reflect(T.ModelType.self),
-                updateBlock: { (view, model) in
-                    (view as! T).updateWithModel(model as! T.ModelType)
-            }))
-        }
-    }
-    
     func registerCellClass<T:ModelTransfer where T: UITableViewCell>(cellClass : T.Type)
     {
-        let reuseIdentifier = RuntimeHelper.classNameFromReflection(_reflect(cellClass))
+        let reuseIdentifier = String(T)
         if self.tableView.dequeueReusableCellWithIdentifier(reuseIdentifier) == nil
         {
             // Storyboard prototype cell
@@ -74,7 +53,7 @@ class TableViewFactory
                 self.registerNibNamed(reuseIdentifier, forCellClass: T.self)
             }
         }
-        self.addMappingForViewType(.Cell, viewClass: T.self)
+        mappings.addMappingForViewType(.Cell, viewClass: T.self)
     }
     
     func registerNibNamed<T:ModelTransfer where T: UITableViewCell>(nibName : String, forCellClass cellClass: T.Type)
@@ -82,55 +61,55 @@ class TableViewFactory
         assert(UINib.nibExistsWithNibName(nibName, inBundle: bundle), "Register cell nib method should be called only if nib exists")
         
         let nib = UINib(nibName: nibName, bundle: bundle)
-        let reuseIdentifier = RuntimeHelper.classNameFromReflection(_reflect(cellClass))
+        let reuseIdentifier = String(T)
         self.tableView.registerNib(nib, forCellReuseIdentifier: reuseIdentifier)
-        self.addMappingForViewType(.Cell, viewClass: T.self)
+        mappings.addMappingForViewType(.Cell, viewClass: T.self)
     }
     
     func registerNiblessHeaderClass<T:ModelTransfer where T: UIView>(headerClass : T.Type)
     {
-        let reuseIdentifier = RuntimeHelper.classNameFromReflection(_reflect(headerClass))
+        let reuseIdentifier = String(T)
         tableView.registerClass(headerClass, forHeaderFooterViewReuseIdentifier: reuseIdentifier)
-        self.addMappingForViewType(.Header, viewClass: T.self)
+        mappings.addMappingForViewType(.SupplementaryView(kind: DTTableViewElementSectionHeader), viewClass: T.self)
     }
     
     func registerNiblessFooterClass<T:ModelTransfer where T: UIView>(footerClass : T.Type)
     {
-        let reuseIdentifier = RuntimeHelper.classNameFromReflection(_reflect(footerClass))
+        let reuseIdentifier = String(T)
         tableView.registerClass(footerClass, forHeaderFooterViewReuseIdentifier: reuseIdentifier)
-        self.addMappingForViewType(.Footer, viewClass: T.self)
+        mappings.addMappingForViewType(.SupplementaryView(kind: DTTableViewElementSectionFooter), viewClass: T.self)
     }
     
     func registerHeaderClass<T:ModelTransfer where T: UIView>(headerClass : T.Type)
     {
-        self.registerNibNamed(RuntimeHelper.classNameFromReflection(_reflect(headerClass)), forHeaderClass: headerClass)
+        self.registerNibNamed(String(T), forHeaderClass: headerClass)
     }
     
     func registerFooterClass<T:ModelTransfer where T:UIView>(footerClass: T.Type)
     {
-        self.registerNibNamed(RuntimeHelper.classNameFromReflection(_reflect(footerClass)), forFooterClass: footerClass)
+        self.registerNibNamed(String(T), forFooterClass: footerClass)
     }
     
     func registerNibNamed<T:ModelTransfer where T:UIView>(nibName: String, forHeaderClass headerClass: T.Type)
     {
         assert(UINib.nibExistsWithNibName(nibName, inBundle: bundle), "Register header nib method should be called only if nib exists")
-        let reuseIdentifier = RuntimeHelper.classNameFromReflection(_reflect(headerClass))
+        let reuseIdentifier = String(T)
         
         if T.isSubclassOfClass(UITableViewHeaderFooterView.self) {
             self.tableView.registerNib(UINib(nibName: nibName, bundle: bundle), forHeaderFooterViewReuseIdentifier: reuseIdentifier)
         }
-        self.addMappingForViewType(.Header, viewClass: T.self)
+        mappings.addMappingForViewType(.SupplementaryView(kind: DTTableViewElementSectionHeader), viewClass: T.self)
     }
     
     func registerNibNamed<T:ModelTransfer where T:UIView>(nibName: String, forFooterClass footerClass: T.Type)
     {
         assert(UINib.nibExistsWithNibName(nibName, inBundle: bundle), "Register footer nib method should be called only if nib exists")
-        let reuseIdentifier = RuntimeHelper.classNameFromReflection(_reflect(footerClass))
+        let reuseIdentifier = String(T.self)
         
         if T.isSubclassOfClass(UITableViewHeaderFooterView.self) {
             self.tableView.registerNib(UINib(nibName: nibName, bundle: bundle), forHeaderFooterViewReuseIdentifier: reuseIdentifier)
         }
-        self.addMappingForViewType(.Footer, viewClass: T.self)
+        mappings.addMappingForViewType(.SupplementaryView(kind: DTTableViewElementSectionFooter), viewClass: T.self)
     }
     
     func cellForModel(model: Any, atIndexPath indexPath:NSIndexPath) -> UITableViewCell
@@ -139,28 +118,31 @@ class TableViewFactory
             preconditionFailure("Received nil model at indexPath: \(indexPath)")
         }
         
-        let typeMirror = RuntimeHelper.mirrorFromModel(unwrappedModel)
-        if let mapping = self.mappingForViewType(.Cell, modelTypeMirror: typeMirror)
+        if let mapping = mappings.mappingCandidatesForViewType(.Cell, model: unwrappedModel).first
         {
-            let cellClassName = RuntimeHelper.classNameFromReflection(mapping.viewTypeMirror)
+            let cellClassName = String(mapping.viewClass)
             let cell = tableView.dequeueReusableCellWithIdentifier(cellClassName, forIndexPath: indexPath)
             mapping.updateBlock(cell, unwrappedModel)
             return cell
         }
         
-        preconditionFailure("Unable to find cell mappings for type: \(_reflect(typeMirror.valueType).summary)")
+        preconditionFailure("Unable to find cell mappings for type: \(model)")
     }
     
     func headerFooterViewWithMapping(mapping: ViewModelMapping, unwrappedModel: Any) -> UIView?
     {
-        let viewClassName = RuntimeHelper.classNameFromReflection(mapping.viewTypeMirror)
+        let viewClassName = String(mapping.viewClass)
         if let view = self.tableView.dequeueReusableHeaderFooterViewWithIdentifier(viewClassName) {
             mapping.updateBlock(view,unwrappedModel)
             return view
         }
         else {
             let view : UIView?
-            if let type = mapping.viewTypeMirror.value as? UIView.Type {
+            
+            // Unfortunately, Swift 2.1 does not allow casting AnyClass.Type to UIView.Type even conditionally,
+            // that's why we're forced to cast through reflect value
+            let mirror = _reflect(mapping.viewClass)
+            if let type = mirror.value as? UIView.Type  {
                 view = type.dt_loadFromXibInBundle(bundle)
             }
             else {
@@ -180,9 +162,7 @@ class TableViewFactory
         
         precondition(unwrappedModel != nil, "Received nil model for headerFooterViewModel")
         
-        let typeMirror = RuntimeHelper.mirrorFromModel(unwrappedModel!)
-        
-        if let mapping = self.mappingForViewType(type, modelTypeMirror: typeMirror) {
+        if let mapping = mappings.mappingCandidatesForViewType(type, model: unwrappedModel).first {
             return self.headerFooterViewWithMapping(mapping, unwrappedModel: unwrappedModel!)
         }
         
@@ -191,11 +171,11 @@ class TableViewFactory
     
     func headerViewForModel(model: Any) -> UIView?
     {
-        return self.headerFooterViewOfType(.Header, model: model)
+        return self.headerFooterViewOfType(.SupplementaryView(kind: DTTableViewElementSectionHeader), model: model)
     }
     
     func footerViewForModel(model: Any) -> UIView?
     {
-        return self.headerFooterViewOfType(.Footer, model: model)
+        return self.headerFooterViewOfType(.SupplementaryView(kind: DTTableViewElementSectionFooter), model: model)
     }
 }